CB2-3183 Forcing file rename
